added map stuff, background tile and player sprites

+++ src/assets/player_sprites.c
@@ -1,5 +1,5 @@
 #include "player_sprites.h"
 
-const int player_sprite[PLAYER_W * PLAYER_H] = {
+const short player_sprite[PLAYER_W * PLAYER_H] = {
 
 };
